Moved classes from jdnssec/main to jdnssec/test
